requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
pandas
psycopg2-binary
networkx
python-dotenv
gdown
requests
gunicorn
Werkzeug
sentence-transformers
scikit-learn
numpy
